refactor(assignment-runtime): extract local state lifecycle
Move persisted state, event observation, and startup recovery into a same-owner mixin. Split request preparation from operation dispatch so the runtime keeps behavior while reducing peak function risk.
